backend/pyproject.toml
[tool.poetry]
name = "contractqa"
version = "0.6.0"
description = "Web application for navigating contracts effortlessly."
authors = ["Jason Wheeler"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
langchain = "^0.0.166"
python-dotenv = "^1.0.0"
openai = "^0.27.6"
pymupdf = "^1.22.3"
werkzeug = "2.2.2"
fastapi = "^0.95.1"
tiktoken = "^0.4.0"
jsonify = "^0.5"
python-multipart = "^0.0.6"
uvicorn = "^0.22.0"
chromadb = "^0.3.22"
websockets = "^11.0.3"
websocket-client = "^1.5.1"
jinja2 = "^3.1.2"
firebase-admin = "^6.1.0"
sentencepiece = "^0.1.99"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"